agent/requirements.txt
google-genai
google-adk
mcp
httpx
google-cloud-aiplatform
asyncio
bs4
nest-asyncio